services/analyzer-client-python/requirements.txt
aiogram==3.0.0b3
pyrogram
grpcio-tools
python-dotenv
pytz